Break up big test class into use cases
